add project page children
