--- app/controllers/Singular.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App;
-
-use Sober\Controller\Controller;
-
-class Singular extends Controller {
-
-
-
-}
